app/requirements.txt
Fastapi
mysql-connector-python
uvicorn